added support for rotated ALC mosaics, some javascript refactoring
