+++ colab_zirc_dims/czd_utils/czd_utils.py
@@ -33,6 +33,7 @@ __all__ = ['check_url',
            'prediction_to_np',
            'mask_list_to_np',
            'rescale_2d_arr',
+           'rotate_pt',
            'crop_nd_arr',
            'mask_to_3D_arr_size',
            'scancsv_to_dict',
@@ -380,6 +381,32 @@ def rescale_2d_arr(im, nR, nC):
     return np.asarray([[ im[int(nR0 * r / nR)][int(nC0 * c / nC)]
                         for c in range(nC)] for r in range(nR)])
 
+def rotate_pt(pt, theta, center):
+    """Rotate an x-y coordinate point clockwise by theta degrees around x-y
+       coordinate point theta.
+
+    Parameters
+    ----------
+    pt : tuple (x, y)
+        Coordinates of point for rotation.
+    theta : float or int
+        Degrees for clockwise rotation of pt around center.
+    center : tuple (x, y)
+        Coordinates of center for rotation operation.
+
+    Returns
+    -------
+    x_new : float
+        X coordinate of rotated point.
+    y_new : float
+        Y coordinate of rotated point.
+
+    """
+    cos_theta, sin_theta = np.cos(np.radians(theta)), np.sin(np.radians(theta))
+    x_new = (pt[0] - center[0]) * cos_theta + (pt[1] - center[1]) * sin_theta + center[0]
+    y_new = -(pt[0] - center[0]) * sin_theta + (pt[1] - center[1]) * cos_theta + center[1]
+    return (x_new, y_new)
+
 
 # from https://stackoverflow.com/a/50322574
 def crop_nd_arr(img, bounding):
@@ -452,8 +479,7 @@ def scancsv_to_dict(scancsv_path):
     temp_coords_dict = {}
     # scanlist from .scancsv file, loaded as a dict
     each_scanlist = pd.read_csv(scancsv_path, header=0, index_col=False,
-                                squeeze=False, encoding='cp1252'
-                                ).to_dict('list')
+                                encoding='cp1252').to_dict('list')
     added_scans_unchanged = []  # list of scans added to output dictionary
 
     # loops through shotlist, gets coordinates for each scan, \
@@ -475,9 +501,9 @@ def scancsv_to_dict(scancsv_path):
 
 
 def get_Align_center_size(align_file_path):
-    """Gets the data tagged as 'Center' and as 'Size' (microns) from a
-       .Align alignment xml file. Used for scaling and mapping shots to
-       mosaic image(s).
+    """Gets the data tagged as 'Center', 'Size' (microns), and 'Rotation'
+       (degrees, clockwise around center) from a .Align alignment xml file.
+       Used for scaling and mapping shots to mosaic image(s).
 
     Parameters
     ----------
@@ -488,15 +514,14 @@ def get_Align_center_size(align_file_path):
     -------
     list
         A list of data extracted from .Align file. Format:
-            [x_center, y_center, x_size, y_size]
+            [x_center, y_center, x_size, y_size, rotation]
 
     """
 
-    x_center, y_center, x_size, y_size = 0, 0, 0, 0
+    x_center, y_center, x_size, y_size, rotation = 0, 0, 0, 0, 0
     align_tree = ET.parse(align_file_path)
     align_root = align_tree.getroot()
-    #loop through xml file to get image center, size
-    ##UPDATE TO INCLUDE ROTATION IF NEEDED##
+    #loop through xml file to get image center, size, rotation if present
     for eachchild in align_root:
         if eachchild.tag == 'Alignment':
             for each_align_data in eachchild:
@@ -508,7 +533,9 @@ def get_Align_center_size(align_file_path):
                     sizes = [float(data) for data
                              in each_align_data.text.split(',')]
                     x_size, y_size = sizes
-    return [x_center, y_center, x_size, y_size]
+                if each_align_data.tag == 'Rotation':
+                    rotation = float(each_align_data.text)
+    return [x_center, y_center, x_size, y_size, rotation]
 
 def calc_scale_factor(Align_x_y_sizes, mosaic_x_y_sizes):
     """Calculate the scale factor for a mosaic image in microns/pixel
@@ -565,8 +592,8 @@ def load_data_dict(project_dir_string):
     mos_csv_path = os.path.join(project_dir_string, 'mosaic_info.csv')
 
     # loads info csv as dictionary
-    mos_csv_dict = pd.read_csv(mos_csv_path, header=0, index_col=False,
-                               squeeze=False).to_dict('list')
+    mos_csv_dict = pd.read_csv(mos_csv_path, header=0, index_col=False
+                               ).to_dict('list')
 
     if not check_mos_csv_keys(mos_csv_dict):
         print('Incorrect mosaic_info.csv headers: correct and re-save')
